♻️ Refactor progress bar (#982)
* ♻️ Added ProgressBar.tsx

* ♻️ Refactor the progressbar for Clinincal Submission

* ♻️ Turn the ClinicalSubmissionProgressBar into React Component
