+++ components/pages/submission-system/program-clinical-submission/index.tsx
@@ -60,42 +60,7 @@ export default function ProgramClinicalSubmission() {
 
   const { data, loading: loadingClinicalSubmission } = useClinicalSubmissionQuery(programShortName);
 
-  const allDataErrors = React.useMemo(
-    () =>
-      data.clinicalSubmissions.clinicalEntities.reduce<
-        Array<
-          ClinicalSubmissionError & {
-            fileName: string;
-          }
-        >
-      >(
-        (acc, entity) => [
-          ...acc,
-          ...entity.dataErrors.map(err => ({
-            ...err,
-            fileName: entity.batchName,
-          })),
-        ],
-        [],
-      ),
-    [data],
-  );
-
-  const hasDataError = !!allDataErrors.length;
-  const hasSchemaError =
-    !!data.clinicalSubmissions.clinicalEntities.length &&
-    data.clinicalSubmissions.clinicalEntities.some(({ schemaErrors }) => !!schemaErrors.length);
-  const hasSomeEntity = data.clinicalSubmissions.clinicalEntities.some(
-    ({ records }) => !!records.length,
-  );
-  const hasSchemaErrorsAfterMigration = data.clinicalSubmissions.state === 'INVALID_BY_MIGRATION';
-  const isReadyForValidation = hasSomeEntity && !hasSchemaError && !hasSchemaErrorsAfterMigration;
-  const isReadyForSignoff = isReadyForValidation && data.clinicalSubmissions.state === 'VALID';
   const isPendingApproval = data.clinicalSubmissions.state === 'PENDING_APPROVAL';
-  const hasUpdate = data.clinicalSubmissions.clinicalEntities.some(
-    clinicalEntity => !!clinicalEntity.dataUpdates.length,
-  );
-  const isValidated = data.clinicalSubmissions.state !== 'OPEN';
 
   return (
     <>
@@ -109,22 +74,6 @@ export default function ProgramClinicalSubmission() {
             showProgress={!loadingClinicalSubmission}
             isPendingApproval={isPendingApproval}
             submissionVersion={data.clinicalSubmissions.version}
-            progressStates={{
-              upload: isReadyForValidation
-                ? 'success'
-                : hasSchemaError || hasSchemaErrorsAfterMigration
-                ? 'error'
-                : 'disabled',
-              validate:
-                isReadyForSignoff || isPendingApproval
-                  ? 'success'
-                  : isReadyForValidation
-                  ? hasDataError
-                    ? 'error'
-                    : 'pending'
-                  : 'disabled',
-              signOff: isReadyForSignoff ? 'pending' : isPendingApproval ? 'success' : 'disabled',
-            }}
           />
         }
       >
